Removed a ton of msvc warnings revealed with OGRE v1.7 (they removed the warning suppressors in OgrePrerequisites.h). All of them are conversions from one type to another that might be lossy (mostly double to float, please always use "3.7f" instead of "3.7" as constants when using floats).

+++ src/libraries/core/GUIManager.cc
@@ -144,7 +144,7 @@ namespace orxonox
         this->luaState_->doFile("InitialiseGUI.lua");
 
         // Align CEGUI mouse with OIS mouse
-        guiSystem_->injectMousePosition(mousePosition.first, mousePosition.second);
+        guiSystem_->injectMousePosition((float)mousePosition.first, (float)mousePosition.second);
 
         // Hide the mouse cursor unless playing in full screen mode
         if (!bFullScreen)
